requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg[binary,pool]>=3.2.0
databricks-sdk>=0.81.0
openai>=1.52.0
pydantic>=2.0.0
python-multipart>=0.0.9
openpyxl>=3.1.0
